HMS-1832 refactor: clean up database operations
Simplify the database operations in the `UpdateAgent` operation.

First, the handler already "merged" the old and new data, so there
is no need to query (`FindByID`) the target record again.

Second, the model object already has the primary key (`id`) set, so
remove the unnecessary `Where` clause.

Third, in `updateIpaDomain()`, after the (cascading) deletion of the
previous associations, simply use `Create` *without*
`Omit(clause.Associations)`.  This recursively inserts the new or
updated associations and avoids the manual iteration and insertion
of the child server, certificate and location objects.

Unfortunately, I could not find a way to avoid the explicit deletion
of the existing associations.  I tried many permutations but in each
case, the old associations were persisting and resulting in
duplicated and/or obsolete values.