HMS-XXXX draft: enforce fips
WIP It needs more work; it is a poc to enforce fips inside the container,
failing if it were not enabled; I could not check out correctly; based on the
links provided by @abbra many time ago (thanks a lot). Making the change
available if it were helpful.

The idea is if the container has the OPENSSL_FORCE_FIPS_MODE=1 then enable a
build tag to build binaries linking the enforcement (it happens on the stage 1
when building the container image); the resulting image would fail when fips is
not enabled (in theory, not checked).

Interesting information:

- https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/9/html/security_hardening/using-the-system-wide-cryptographic-policies_security-hardening#enabling-fips-mode-in-a-container_using-the-system-wide-cryptographic-policies
- https://developers.redhat.com/articles/2022/05/31/your-go-application-fips-compliant

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>