fix: update ref to api
Update the reference of the api/ directory to the most recent one,
and regenerate api boilerplate code.

The changes in the generated code does not impact, because the value
added to the slice in the context is not used currently in the backend
source code.

(not yet) but an enhancement on the future could be simplify the
middleware for XRHID enforcement based on this values (credits
to @cheimes who raised this idea initially).

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>